# public/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "GLAP API",
    "description": "GLAP API documentation"
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            }
          }
        }
      }
    },
    "/auth/signin": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "user": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "632e1c7a56c75874e0c6a1233"
          },
          "name": {
            "type": "string",
            "example": "jhon Doe"
          },
          "identification_number": {
            "type": "string",
            "example": "1234567891"
          },
          "email": {
            "type": "string",
            "example": "jhon.doe@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "jhon"
          },
          "role": {
            "type": "object",
            "properties": {
              "0": {
                "type": "string",
                "example": "enterprise"
              },
              "1": {
                "type": "string",
                "example": "person"
              },
              "enterprise": {
                "type": "number",
                "example": 0
              },
              "person": {
                "type": "number",
                "example": 1
              }
            }
          },
          "provider_type": {
            "type": "object",
            "properties": {
              "0": {
                "type": "string",
                "example": "gas"
              },
              "1": {
                "type": "string",
                "example": "electricity"
              },
              "2": {
                "type": "string",
                "example": "cliff"
              },
              "gas": {
                "type": "number",
                "example": 0
              },
              "electricity": {
                "type": "number",
                "example": 1
              },
              "cliff": {
                "type": "number",
                "example": 2
              }
            }
          }
        },
        "xml": {
          "name": "user"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}